random world generator, generates random paths for now, town tile and tree chunks are now on a transparent decoration matrix (no more white background), world map is now beeing saved in a json file instead of hardcoded(towns have to follow). for testing the random world generator so far go to class "ScreensController" and switch line 137 with 136. For generating the world map use the main in class Test, it will create the world map field once and save it on the harddrive.

+++ src/main/java/org/csproject/model/bean/Town.java
@@ -7,6 +7,8 @@ import java.util.Map;
 
 /**
  * Created by Brett on 10/19/2015.
+ *
+ * TODO Redundant: A town is a field
  */
 public class Town {
     private Map<String, NavigationPoint> startPoints;
@@ -14,7 +16,8 @@ public class Town {
 
     public Town() {
         startPoints = new HashMap<>();
-        startPoints.put("townEntrance", new NavigationPoint(18 * Constants.TILE_SIZE, 18 * Constants.TILE_SIZE));
+        startPoints.put("townEntrance",
+                new NavigationPoint(18 * (int)Constants.TILE_SIZE, 18 * (int)Constants.TILE_SIZE));
     }
 
     public NavigationPoint getTownEntrance() {
